reworked almost everything. some stuff doesnt work right now

+++ src/component_orb.hpp
@@ -4,10 +4,10 @@
 #include "entityx/entityx.h"
 
 struct Orb : entityx::Component<Orb>{
-    Orb(float score = 1, float max_lifetime = 0.f) : m_score(score), m_max_lifetime(max_lifetime), m_current_lifetime(0.f) {
+    Orb(unsigned int score = 1, float max_lifetime = 0.f) : m_score(score), m_max_lifetime(max_lifetime), m_current_lifetime(0.f) {
     }
 
-    float score() {
+    unsigned int score() {
         return m_score;
     }
 
@@ -22,7 +22,7 @@ struct Orb : entityx::Component<Orb>{
     }
 
   private:
-    float m_score;
+    unsigned int m_score;
     float m_max_lifetime;
     float m_current_lifetime;
 };
